Proyecto completo

+++ src/main/java/org/bedu/postwork/controllers/ProductoController.java
@@ -1,8 +1,7 @@
 package org.bedu.postwork.controllers;
 
 import lombok.RequiredArgsConstructor;
-import org.bedu.postwork.model.Cliente;
-import org.bedu.postwork.model.Producto;
+import org.bedu.postwork.model.ProductoModel;
 import org.bedu.postwork.services.ProductoService;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -10,7 +9,6 @@ import org.springframework.web.bind.annotation.*;
 import org.springframework.web.server.ResponseStatusException;
 
 import java.net.URI;
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
@@ -22,9 +20,9 @@ public class ProductoController {
     private final ProductoService productoService;
 
     @GetMapping("/{productoId}")
-    public ResponseEntity<Producto> getProducto(@PathVariable Long productoId)
+    public ResponseEntity<ProductoModel> getProducto(@PathVariable Long productoId)
     {
-        Optional<Producto> productoDb = productoService.obtenerProducto(productoId);
+        Optional<ProductoModel> productoDb = productoService.obtenerProducto(productoId);
 
         if (productoDb.isEmpty()) {
             throw new ResponseStatusException(HttpStatus.NOT_FOUND, "El producto especificado no existe.");
@@ -34,9 +32,9 @@ public class ProductoController {
     }
 
     @GetMapping
-    public  ResponseEntity<List<Producto>> getProductos()
+    public  ResponseEntity<List<ProductoModel>> getProductos()
     {
-        List<Producto> productosDb = productoService.listarProductos();
+        List<ProductoModel> productosDb = productoService.listarProductos();
 
         if(productosDb.isEmpty())
         {
@@ -46,25 +44,40 @@ public class ProductoController {
     }
 
     @PostMapping
-    public ResponseEntity<Void> crearProducto(@RequestBody Producto producto)
+    public ResponseEntity<Void> crearProducto(@RequestBody ProductoModel productoModel)
     {
-        Producto productoNuevo = productoService.guardarProducto(producto);
+        ProductoModel productoModelNuevo = productoService.guardarProducto(productoModel);
 
-        return ResponseEntity.created(URI.create(String.valueOf(productoNuevo.getId()))).build();
+        return ResponseEntity.created(URI.create(String.valueOf(productoModelNuevo.getId()))).build();
     }
 
     @PutMapping("/{productoId}")
-    public ResponseEntity<Void> actualizarProducto(@PathVariable Long productoId, @RequestBody Producto producto)
+    public ResponseEntity<Void> actualizarProducto(@PathVariable Long productoId, @RequestBody ProductoModel productoModel)
     {
-        productoService.actualizarProducto(producto);
-        return ResponseEntity.status(HttpStatus.OK).build();
+        if(productoService.obtenerProducto(productoId).get() == null)
+        {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "El producto especificado no existe.");
+        }
+        else
+        {
+            productoModel.setId(productoId);
+            productoService.actualizarProducto(productoModel);
+            return ResponseEntity.status(HttpStatus.OK).build();
+        }
     }
 
     @DeleteMapping("/{productoId}")
     public  ResponseEntity<Void> eliminarProducto(@PathVariable Long productoId)
     {
-        productoService.eliminarProducto(productoId);
-        return ResponseEntity.status(HttpStatus.OK).build();
+        if(productoService.obtenerProducto(productoId).get() == null)
+        {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "El producto especificado no existe.");
+        }
+        else
+        {
+            productoService.eliminarProducto(productoId);
+            return ResponseEntity.status(HttpStatus.OK).build();
+        }
     }
 
     @GetMapping("/numProductos")
